+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file[^1].
 ## [Unreleased]
 ### Added
 - filter catalog by curator and fix harvesting of curator data
+- attributes "format", "deliveryPoint" and "note" to JIRA request in order process
 
 ### Fixed
 - calculation of isFree for artworks with authorities without death_year